requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
supabase==1.2.0
pydantic==2.6.3
httpx==0.27.0
python-multipart==0.0.9
tweepy==4.14.0 